+++ src/matching/matching_interface.jl
@@ -14,7 +14,6 @@ Supertype of all "matchers" that match can IDs labelling attractors.
 Currently available matchers:
 
 - [`MatchBySSSetDistance`](@ref)
-- [`MatchByFeatureDistance`](@ref)
 - [`MatchByBasinEnclosure`](@ref)
 - [`MatchByBasinOverlap`](@ref)
 - [`DontMatch`](@ref)
@@ -224,4 +223,3 @@ _use_vanished(matcher) = false
 include("basin_overlap.jl")
 include("sssdistance.jl")
 include("basin_enclosure.jl")
-include("feature_distance.jl")